nba_dash_v2/requirements.txt
python-dotenv
rpy2 --pre
mysql-connector-python
pandas
numpy
nba_api
pymysql
config